# Optimize key creation routines.