--- a/testutils/keys.go
+++ b/testutils/keys.go
@@ -1,0 +1,25 @@
+// Copyright 2016 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+// Author: Peter Mattis (peter@cockroachlabs.com)
+
+package testutils
+
+import "bytes"
+
+// MakeKey makes a new key which is the concatenation of the given inputs, in
+// order.
+func MakeKey(keys ...[]byte) []byte {
+	return bytes.Join(keys, nil)
+}