… keys/printer_test.go