Remove usage of {keys,roachpb}.MakeKey from common paths. Moved test to
testutils.MakeKey to prevent accidental reintroduction. The key creation
routines in `keys` now preallocate a buffer of the correct size.

```
name                   old time/op    new time/op    delta
Insert1_Cockroach-8       379µs ± 2%     370µs ± 0%  -2.55%   (p=0.000 n=10+9)
Insert10_Cockroach-8      775µs ± 2%     751µs ± 2%  -3.05%  (p=0.000 n=10+10)
Insert100_Cockroach-8    4.17ms ± 1%    4.07ms ± 0%  -2.22%   (p=0.000 n=10+9)
Update1_Cockroach-8       857µs ± 1%     843µs ± 1%  -1.64%  (p=0.000 n=10+10)
Update10_Cockroach-8     4.65ms ± 0%    4.59ms ± 0%  -1.31%  (p=0.000 n=10+10)
Update100_Cockroach-8    41.0ms ± 1%    40.3ms ± 1%  -1.69%  (p=0.000 n=10+10)
Delete1_Cockroach-8      1.18ms ± 0%    1.18ms ± 1%    ~      (p=0.887 n=7+10)
Delete10_Cockroach-8     1.93ms ± 1%    1.90ms ± 1%  -1.49%  (p=0.001 n=10+10)
Delete100_Cockroach-8    11.0ms ± 2%    10.7ms ± 0%  -3.21%    (p=0.000 n=8+8)
Scan1_Cockroach-8         181µs ± 1%     180µs ± 0%  -0.55%   (p=0.028 n=10+9)
Scan10_Cockroach-8        223µs ± 1%     221µs ± 1%  -0.87%  (p=0.001 n=10+10)
Scan100_Cockroach-8       614µs ± 2%     597µs ± 3%  -2.78%  (p=0.019 n=10+10)

name                   old allocs/op  new allocs/op  delta
Insert1_Cockroach-8         474 ± 0%       438 ± 0%  -7.59%   (p=0.000 n=9+10)
Insert10_Cockroach-8      1.12k ± 0%     1.08k ± 0%  -3.28%  (p=0.000 n=10+10)
Insert100_Cockroach-8     7.31k ± 0%     7.27k ± 0%  -0.52%   (p=0.000 n=9+10)
Update1_Cockroach-8       1.06k ± 0%     0.99k ± 0%  -6.50%  (p=0.000 n=10+10)
Update10_Cockroach-8      7.11k ± 0%     6.75k ± 0%  -5.15%    (p=0.000 n=8+7)
Update100_Cockroach-8     67.5k ± 0%     64.1k ± 0%  -4.95%   (p=0.000 n=10+9)
Delete1_Cockroach-8       1.01k ± 0%     0.95k ± 0%  -6.46%  (p=0.000 n=10+10)
Delete10_Cockroach-8      2.49k ± 0%     2.42k ± 0%  -2.67%   (p=0.000 n=10+9)
Delete100_Cockroach-8     16.6k ± 0%     16.5k ± 0%  -0.39%   (p=0.000 n=10+8)
Scan1_Cockroach-8           285 ± 0%       280 ± 0%  -1.86%  (p=0.000 n=10+10)
Scan10_Cockroach-8          376 ± 0%       370 ± 0%  -1.46%   (p=0.000 n=10+9)
Scan100_Cockroach-8       1.11k ± 0%     1.11k ± 0%  -0.42%   (p=0.000 n=10+8)
```

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/4173)
<!-- Reviewable:end -->