@tamird 2016-02-06
Reviewed 5 of 15 files at r1.
Review status: 5 of 15 files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[keys/keys.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9o9ocslL8wlsChjvLC:-K9o9ocslL8wlsChjvLD:-392313706) ([raw file](https://github.com/cockroachdb/cockroach/blob/89dd4bb4202712715403ea469a430349c85bd3dd/keys/keys.go#L33)):*
Isn't this all just because `roachpb.MakeKey` is strangely inefficient?

Isn't it enough to apply this diff:
```diff
diff --git a/roachpb/data.go b/roachpb/data.go
index 723faa7..86664ab 100644
--- a/roachpb/data.go
+++ b/roachpb/data.go
@@ -112,11 +112,7 @@ type Key []byte
 // MakeKey makes a new key which is the concatenation of the
 // given inputs, in order.
 func MakeKey(keys ...[]byte) []byte {
-	byteSlices := make([][]byte, len(keys))
-	for i, k := range keys {
-		byteSlices[i] = []byte(k)
-	}
-	return bytes.Join(byteSlices, nil)
+	return bytes.Join(keys, nil)
 }
 
 // BytesNext returns the next possible byte by appending an \x00.

```

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173)*
<!-- Sent from Reviewable.io -->