@petermattis 2016-02-06
Review status: 5 of 15 files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[keys/keys.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9o9ocslL8wlsChjvLC:-K9oAIXbVddzwg8GUou-:-1075757096) ([raw file](https://github.com/cockroachdb/cockroach/blob/89dd4bb4202712715403ea469a430349c85bd3dd/keys/keys.go#L33)):*
No, that gets part of the way there, but creating a single buffer of the correct length is better when we toss in the `encoding.Encode*` routines as well.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173)*
<!-- Sent from Reviewable.io -->