@tamird 2016-02-06
Review status: 5 of 15 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[keys/keys.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9o9ocslL8wlsChjvLC:-K9oDeekgdYApSnkKciM:-170673961) ([raw file](https://github.com/cockroachdb/cockroach/blob/89dd4bb4202712715403ea469a430349c85bd3dd/keys/keys.go#L33)):*
it's really hard to read this diff - there are a bunch of changes lumped into one commit: unexporting makeKey, moving stuff to testutils. can you break things up?

---

*[keys/keys.go, line 37 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9oDztl0kr5UtQrz9Q7:-K9oDztl0kr5UtQrz9Q8:-234895084) ([raw file](https://github.com/cockroachdb/cockroach/blob/4c875ca75d04c617e1c21bc77cacdeb349510a2d/keys/keys.go#L37)):*
i dont know what i'm missing here - this looks exactly the same as bytes.Join

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173)*
<!-- Sent from Reviewable.io -->