@petermattis 2016-02-06
Review status: 5 of 15 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[keys/keys.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9o9ocslL8wlsChjvLC:-K9oEuhnXBf6hdXY9-G5:-40219114) ([raw file](https://github.com/cockroachdb/cockroach/blob/89dd4bb4202712715403ea469a430349c85bd3dd/keys/keys.go#L33)):*
The changes are all related and, outside of this file, mostly mechanical. I'm not aware of any requirement to place potentially separable changes in separate commits. Certainly, some people do that, but that's their prerogative. 

I just took another scan through the diff and it still looks mostly mechanical to me (again, outside of  this file). Perhaps take another look on Monday with fresher eyes.

---

*[keys/keys.go, line 37 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9oDztl0kr5UtQrz9Q7:-K9oET14D2EzV8pQUnh2:1196025850) ([raw file](https://github.com/cockroachdb/cockroach/blob/4c875ca75d04c617e1c21bc77cacdeb349510a2d/keys/keys.go#L37)):*
`bytes.Join` requires passing a `[][]byte`. Where did that come from? And this is the simplest case. See `StoreStatusKey` for an example where `encoding.EncodeUvarintAscending` is used which has to allocate if it isn't passed an appropriately sized buffer.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173)*
<!-- Sent from Reviewable.io -->