@tamird 2016-02-06
LGTM

---

Reviewed 10 of 15 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[keys/keys.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-K9o9ocslL8wlsChjvLC:-K9oIn0O2Xiuwoxj2Gyr:-665118321) ([raw file](https://github.com/cockroachdb/cockroach/blob/89dd4bb4202712715403ea469a430349c85bd3dd/keys/keys.go#L33)):*
There's no requirement, it's just hard to review because I can't skip over all the mechanical changes.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4173#-:-K9oJXJ4Ryxelp_OIFla:676834233)*
<!-- Sent from Reviewable.io -->